Add files via upload
This MATLAB code uses the Metropolis algorithm (for sequential updates) and simulated annealing (to ensure global optimization) to simulate the process of protein folding. Proteins are modelled as chains of two kinds of mutually repulsive amino acids -- hydrophilic and hydrophobic -- that localize themselves in a plane in such a way that the potential energy is minimized.